[3party][succinct] Fixed uint64_t <-> size_t warnings. - Use size_t for memory byte index - Use uint64_t for bit position index.

+++ 3party/succinct/elias_fano_compressed_list.hpp
@@ -47,7 +47,7 @@ namespace succinct {
             return ((uint64_t(1) << l) | m_bits.get_bits(r.first, l)) - 1;
         }
 
-        size_t size() const
+        uint64_t size() const
         {
             return m_ef.num_ones() - 1;
         }
